+++ siw-food/src/main/java/it/uniroma3/siwfood/controller/RicettaController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import it.uniroma3.siwfood.model.Cuoco;
 import it.uniroma3.siwfood.model.Ricetta;
+import it.uniroma3.siwfood.model.auth.User;
 import it.uniroma3.siwfood.service.CuocoService;
 import it.uniroma3.siwfood.service.RicettaService;
 import it.uniroma3.siwfood.service.auth.UserService;
@@ -61,31 +61,38 @@ public class RicettaController extends GlobalController {
 
 
 
-    @GetMapping("/cuoco/editRicetta/{ricetta_id}/{user_id}")
-    public String formModificaRicettaDalCuoco(@PathVariable("ricetta_id") Long ricetta_id, @PathVariable("user_id") Long user_id, Model model) {
-        Cuoco cuoco = this.userService.findById(user_id).getCuoco();
-        Ricetta ricetta = this.ricettaService.findById(ricetta_id);
-        if(ricetta.getCuoco().equals(cuoco)) {
-            model.addAttribute("ricetta", ricetta);
-            return "forms/formModificaRicettaDalCuoco.html";
-        }
-        else return "redirect:/ricette" + ricetta_id;
-    }
-    @PostMapping("/cuoco/editRicetta/{ricetta_id}")
-    public String editRicettaCuoco(@PathVariable("ricetta_id") Long id, @ModelAttribute Ricetta ricetta) {
-        ricetta.setId(id);
-        this.ricettaService.save(ricetta);
-        return "redirect:/ricette/" + id;
-    }
+//    @GetMapping("/cuoco/editRicetta/{ricetta_id}/{user_id}")
+//    public String formModificaRicettaDalCuoco(@PathVariable("ricetta_id") Long ricetta_id, @PathVariable("user_id") Long user_id, Model model) {
+//        Cuoco cuoco = this.userService.findById(user_id).getCuoco();
+//        Ricetta ricetta = this.ricettaService.findById(ricetta_id);
+//        if(ricetta.getCuoco().equals(cuoco)) {
+//            model.addAttribute("ricetta", ricetta);
+//            return "forms/formModificaRicettaDalCuoco.html";
+//        }
+//        else return "redirect:/ricette" + ricetta_id;
+//    }
+//    @PostMapping("/cuoco/editRicetta/{ricetta_id}")
+//    public String editRicettaCuoco(@PathVariable("ricetta_id") Long id, @ModelAttribute Ricetta ricetta) {
+//        ricetta.setId(id);
+//        this.ricettaService.save(ricetta);
+//        return "redirect:/ricette/" + id;
+//    }
 
 
 
     @GetMapping("/admin/editRicetta/{ricetta_id}")
     public String editRicettaAdmin(@PathVariable("ricetta_id") Long id, Model model) {
-        model.addAttribute("ricetta", this.ricettaService.findById(id));
-        return "forms/formModificaRicettaAdmin.html";
+        User user = getCredentials().getUser();
+        if (getCredentials().isAdmin()
+        && cuocoService.findByNomeAndCognome(user.getNome(), user.getCognome()).getId() != id) {
+            model.addAttribute("ricetta", this.ricettaService.findById(id));
+            return "forms/formModificaRicettaAdmin.html";
+        }
+        else {
+            model.addAttribute("messaggioErrore", "Non disponi per le autorizzazioni necessarie per questa operazione!");
+            return "messaggioErrore";
+        }
     }
-
     @PostMapping("/admin/editRicetta/{ricetta_id}")
     public String editRicettaAdmin(@PathVariable("ricetta_id") Long id, @ModelAttribute Ricetta ricetta) {
         ricetta.setId(id);
@@ -95,52 +102,67 @@ public class RicettaController extends GlobalController {
 
 
 
-    @GetMapping("cuoco/deleteRicetta/{ricetta_id}/{user_id}")
-    public String deleteRicettaCuoco(@PathVariable("ricetta_id") Long ricetta_id, @PathVariable("user_id") Long user_id) {
-        Ricetta ricetta = this.ricettaService.findById(ricetta_id);
-        Cuoco cuoco = this.userService.findById(user_id).getCuoco();
-        if(ricetta.getCuoco().equals(cuoco)) {
-            this.ricettaService.deleteById(ricetta_id);
-            return "redirect:/ricette";
-        }
-        else return "redirect:/ricette/" + ricetta_id;
-    }
+//    @GetMapping("cuoco/deleteRicetta/{ricetta_id}/{user_id}")
+//    public String deleteRicettaCuoco(@PathVariable("ricetta_id") Long ricetta_id, @PathVariable("user_id") Long user_id) {
+//        Ricetta ricetta = this.ricettaService.findById(ricetta_id);
+//        Cuoco cuoco = this.userService.findById(user_id).getCuoco();
+//        if(ricetta.getCuoco().equals(cuoco)) {
+//            this.ricettaService.deleteById(ricetta_id);
+//            return "redirect:/ricette";
+//        }
+//        else return "redirect:/ricette/" + ricetta_id;
+//    }
 
 
     //DAL SISTEMA
     @GetMapping("/admin/deleteRicetta/{ricetta_id}")
     public String deleteRicettaAdmin(@PathVariable("ricetta_id") Long id) {
-        this.ricettaService.deleteById(id);
-        return "redirect:/ricette";
-    }
-
-
-
-    @GetMapping("/cuoco/addRicetta/{cuoco_id}/{user_id}")
-    public String formNuovaRicettaCuoco(@PathVariable("cuoco_id") Long idC, @PathVariable("user_id") Long idU,Model model) {
-        
-        if(!(this.cuocoService.findById(idC).equals(this.userService.findById(idU).getCuoco()))){
-            return "redirect:/cuochi/" + idC;
+        User user = getCredentials().getUser();
+        if (getCredentials().isAdmin()
+        && cuocoService.findByNomeAndCognome(user.getNome(), user.getCognome()).getId() != id) {
+            this.ricettaService.deleteById(id);
+            return "redirect:/ricette";
         }
-        
-        model.addAttribute("cuoco", this.cuocoService.findById(idC));
-        model.addAttribute("ricetta", new Ricetta());
-        return "forms/formNuovaRicettaDalCuoco.html";
+        else {
+            return "redirect:/error";
+        }
     }
-    @PostMapping("/cuochi/addRicetta/{cuoco_id}")
-    public String addRicettaCuoco(@PathVariable("cuoco_id") Long cuocoId, @ModelAttribute Ricetta ricetta) {
-        ricetta.setCuoco(this.cuocoService.findById(cuocoId));
-        this.ricettaService.save(ricetta);
-        return "redirect:/cuochi/" + cuocoId;
-    }
+
+
+
+//    @GetMapping("/cuoco/addRicetta/{cuoco_id}/{user_id}")
+//    public String formNuovaRicettaCuoco(@PathVariable("cuoco_id") Long idC, @PathVariable("user_id") Long idU,Model model) {
+//        
+//        if(!(this.cuocoService.findById(idC).equals(this.userService.findById(idU).getCuoco()))){
+//            return "redirect:/cuochi/" + idC;
+//        }
+//        
+//        model.addAttribute("cuoco", this.cuocoService.findById(idC));
+//        model.addAttribute("ricetta", new Ricetta());
+//        return "forms/formNuovaRicettaDalCuoco.html";
+//    }
+//    @PostMapping("/cuochi/addRicetta/{cuoco_id}")
+//    public String addRicettaCuoco(@PathVariable("cuoco_id") Long cuocoId, @ModelAttribute Ricetta ricetta) {
+//        ricetta.setCuoco(this.cuocoService.findById(cuocoId));
+//        this.ricettaService.save(ricetta);
+//        return "redirect:/cuochi/" + cuocoId;
+//    }
 
 
 
     @GetMapping("/admin/addRicetta/{cuoco_id}")
     public String formNuovaRicettaAdmin(@PathVariable("cuoco_id") Long id,Model model) {
-        model.addAttribute("cuoco", this.cuocoService.findById(id));
-        model.addAttribute("ricetta", new Ricetta());
-        return "forms/formNuovaRicettaAdmin.html";
+        User user = getCredentials().getUser();
+        if (getCredentials().isAdmin()
+        && cuocoService.findByNomeAndCognome(user.getNome(), user.getCognome()).getId() != id) {
+            model.addAttribute("cuoco", this.cuocoService.findById(id));
+            model.addAttribute("ricetta", new Ricetta());
+            return "forms/formNuovaRicettaAdmin.html";
+        }
+        else {
+            model.addAttribute("messaggioErrore", "Non disponi per le autorizzazioni necessarie per questa operazione!");
+            return "messaggioErrore";
+        }
     }
     @PostMapping("/admin/addRicetta/{cuoco_id}")
     public String addRicettaAdmin(@PathVariable("cuoco_id") Long id, @ModelAttribute Ricetta ricetta) {
